code (1)/backend/requirements.txt
# Smart Invoice Assistant - Backend Dependencies

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
openai==1.54.0
supabase==2.9.1
python-dotenv==1.0.1
pydantic==2.9.2
